requirements.txt
python-dotenv==1.0.0
urllib3==2.0.4
discord.py==2.3.2
